Move Browser e2e build
